PR #996: Fixing PVS-Studio warnings (Contributed by PaulEremeeff)
I have reviewed each change carefully, but it is a large change and I could have missed something! Here is a summary of the types of changes in this CL:
 * Made nullptr checks consistent (the plurality of the changes are of this type)
 * Completed switch statements (IE, switch did not explicitly handle default case, but had unhandled enum entries - this is the second most popular type of fix)
 * Removed unused variables
 * Removed redundant initializations
 * WidgetNavigationCustomization.cpp was fixed by the owner
 * integers converted to floats where result was stored in a float
 * Removed redundent null checks (e.g. before delete statements)
 * Renamed variables to prevent non-obvious shadowing
 * Fixed use of bitwise & when checking for equality to an enum entry (which is often 0)
 * Fixes for some copy paste errors (e.g. FoliageEdMode.cpp)

[CL 2498053 by Dan Oconnor in Main branch]

+++ Engine/Source/Runtime/Landscape/Private/LandscapeGizmoActor.cpp
@@ -828,8 +828,8 @@ void ALandscapeGizmoActiveActor::SampleData(int32 SizeX, int32 SizeY)
 		{
 			for (int32 X = 0; X < TexSizeX; ++X)
 			{
-				float TexX = X * SizeX / TexSizeX;
-				float TexY = Y * SizeY / TexSizeY;
+				float TexX = static_cast<float>(X) * SizeX / TexSizeX;
+				float TexY = static_cast<float>(Y) * SizeY / TexSizeY;
 				int32 LX = FMath::FloorToInt(TexX);
 				int32 LY = FMath::FloorToInt(TexY);
